Code/assignment1/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=0.15.0=pypi_0
anyio=3.7.1=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.1.0=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
bleach=6.0.0=pypi_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.2.0=pypi_0
clang=5.0=pypi_0
comm=0.1.3=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
exceptiongroup=1.1.2=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.18.0=pypi_0
flatbuffers=1.12=pypi_0
fqdn=1.5.1=pypi_0
gast=0.4.0=pypi_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.56.2=pypi_0
h5py=3.1.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.0=pypi_0
ipykernel=6.25.0=pypi_0
ipython=8.12.2=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.7=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.1=pypi_0
json5=0.9.14=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.18.4=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.3.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.3.1=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-lsp=2.2.0=pypi_0
jupyter-server=2.7.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=4.0.3=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.24.0=pypi_0
jupyterlab-widgets=3.0.8=pypi_0
keras=2.6.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.1=pypi_0
nbclient=0.8.0=pypi_0
nbconvert=7.7.3=pypi_0
nbformat=5.9.1=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.7=pypi_0
networkx=3.1=pypi_0
notebook=7.0.0=pypi_0
notebook-shim=0.2.3=pypi_0
numpy=1.21.0=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.0.74=pypi_0
openssl=3.0.9=h7f8727e_0
opt-einsum=3.3.0=pypi_0
overrides=7.3.1=pypi_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.0.0=pypi_0
pip=23.2.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.9.1=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pygments=2.15.1=pypi_0
python=3.8.17=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
pytz=2023.3=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=pypi_0
qtconsole=5.4.3=pypi_0
qtpy=2.3.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.0=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.9.2=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.10.1=pypi_0
send2trash=1.8.2=pypi_0
setuptools=68.0.0=py38h06a4308_0
six=1.15.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.2=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow-addons=0.21.0=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-gpu=2.6.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
tornado=6.3.2=pypi_0
traitlets=5.9.0=pypi_0
typeguard=2.13.3=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023.3=pypi_0
uri-template=1.3.0=pypi_0
urllib3=1.26.16=pypi_0
visdom=0.2.4=pypi_0
wcwidth=0.2.6=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.1=pypi_0
werkzeug=2.3.6=pypi_0
wheel=0.38.4=py38h06a4308_0
widgetsnbextension=4.0.8=pypi_0
wrapt=1.12.1=pypi_0
xz=5.4.2=h5eee18b_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
